Extract blocked issue recovery reconciliation
