+++ src/recovery-reconciliation.ts
@@ -13,25 +13,17 @@ import {
   isOpenPullRequest,
 } from "./supervisor/supervisor-lifecycle";
 import { inferFailureContext } from "./supervisor/supervisor-failure-context";
-import { shouldReconcileTrackedPrStaleReviewBot } from "./supervisor/supervisor-execution-policy";
-import {
-  findHighRiskBlockingAmbiguity,
-  hasAvailableIssueLabels,
-  lintExecutionReadyIssueBody,
-} from "./issue-metadata";
 import { buildIssueDefinitionFingerprint, issueDefinitionFreshnessPatch } from "./issue-definition-freshness";
 import { RecoveryEvent } from "./run-once-cycle-prelude";
 import { StateStore } from "./core/state-store";
-import { GitHubIssue, GitHubPullRequest, IssueRunRecord, PullRequestCheck, ReviewThread, RunState, SupervisorConfig, SupervisorStateFile } from "./core/types";
+import { GitHubIssue, IssueRunRecord, PullRequestCheck, ReviewThread, RunState, SupervisorConfig, SupervisorStateFile } from "./core/types";
 import { nowIso, truncate } from "./core/utils";
-import { resetTrackedPrHeadScopedStateOnAdvance } from "./tracked-pr-lifecycle-projection";
 import {
   buildSupervisorMutationRecordSnapshot,
   type SupervisorMutationRecordSnapshotDto,
   type SupervisorMutationResultDto,
 } from "./supervisor/supervisor-mutation-report";
 import { STALE_STABILIZING_NO_PR_RECOVERY_SIGNATURE } from "./no-pull-request-state";
-import { applyFailureSignature } from "./supervisor/supervisor-failure-helpers";
 import {
   doneResetPatch,
   sanitizeRecoveryReason,
@@ -44,11 +36,10 @@ import {
 } from "./recovery-historical-reconciliation";
 import { reconcileParentEpicClosuresInModule } from "./recovery-parent-epic-reconciliation";
 import {
-  buildTrackedPrResumeRecoveryEvent,
   reconcileStaleFailedTrackedPrRecord,
   reconcileTrackedMergedButOpenIssuesInModule,
-  suppressSameHeadNoProgressReviewThreadRecovery,
 } from "./recovery-tracked-pr-reconciliation";
+import { reconcileRecoverableBlockedIssueStatesInModule } from "./recovery-blocked-issue-reconciliation";
 export {
   inspectOrphanedWorkspacePruneCandidates,
   pruneOrphanedWorkspacesForOperator,
@@ -56,20 +47,6 @@ export {
   type OrphanedWorkspacePruneEligibility,
 } from "./recovery-workspace-reconciliation";
 import { cleanupExpiredDoneWorkspaces as cleanupExpiredDoneWorkspacesInModule } from "./recovery-workspace-reconciliation";
-import { buildTrackedPrStaleFailureConvergencePatch } from "./recovery-tracked-pr-support";
-import { mergeConflictDetected, summarizeChecks } from "./supervisor/supervisor-status-rendering";
-import { projectTrackedPrLifecycle } from "./tracked-pr-lifecycle-projection";
-import { hasFreshTrackedPrReadyPromotionBlockerEvidence } from "./tracked-pr-ready-promotion-blocker";
-import { queuedReadyPromotionPathHygieneRepairContext } from "./ready-promotion-path-hygiene-repair";
-import { clearRequirementsBlockerIssueComment } from "./requirements-blocker-issue-comment";
-import { syncTrackedPrPersistentStatusComment } from "./tracked-pr-status-comment";
-import {
-  configuredBotReviewThreads,
-  manualReviewThreads,
-  latestReviewComment,
-  latestReviewCommentAuthorIsAllowedBot,
-} from "./review-thread-reporting";
-import { hasCodexConnectorFindingReviewComment } from "./codex-connector-review-policy";
 
 const OPERATOR_REQUEUEABLE_STATES = new Set<RunState>(["blocked", "failed"]);
 type StaleStabilizingNoPrBranchState = "recoverable" | "already_satisfied_on_main";
@@ -296,71 +273,6 @@ function shouldReconsiderGenericNoPrIssueDefinitionChange(
   }
 
   return buildIssueDefinitionFingerprint(issue) !== record.issue_definition_fingerprint;
-}
-
-function firstPassingCheckEvidence(checks: PullRequestCheck[]): string | null {
-  if (checks.length === 0) {
-    return null;
-  }
-
-  const checkSummary = summarizeChecks(checks);
-  if (!checkSummary.allPassing || checkSummary.hasPending || checkSummary.hasFailing) {
-    return null;
-  }
-
-  const firstCheckName = checks
-    .map((check) => check.name.trim())
-    .filter((name) => name.length > 0)
-    .sort()[0];
-
-  return `required_checks_green:${firstCheckName ?? "all"}`;
-}
-
-function currentHeadConfiguredBotEvidence(
-  pr: Pick<
-    GitHubPullRequest,
-    | "configuredBotCurrentHeadObservedAt"
-    | "configuredBotCurrentHeadStatusState"
-    | "configuredBotTopLevelReviewStrength"
-  >,
-): string | null {
-  if (!pr.configuredBotCurrentHeadObservedAt) {
-    return null;
-  }
-
-  const statusState = pr.configuredBotCurrentHeadStatusState?.toLowerCase() ?? null;
-  const statusPassed =
-    statusState === "success" ||
-    statusState === "pass" ||
-    statusState === "passed";
-  const topLevelPassed =
-    pr.configuredBotTopLevelReviewStrength === null ||
-    pr.configuredBotTopLevelReviewStrength === undefined ||
-    pr.configuredBotTopLevelReviewStrength === "nitpick_only";
-
-  if (!statusPassed && !topLevelPassed) {
-    return null;
-  }
-
-  return "configured_bot_current_head_passed";
-}
-
-function trackedHandoffExternalProgressEvidence(args: {
-  record: Pick<IssueRunRecord, "last_head_sha">;
-  pr: Pick<
-    GitHubPullRequest,
-    | "configuredBotCurrentHeadObservedAt"
-    | "configuredBotCurrentHeadStatusState"
-    | "configuredBotTopLevelReviewStrength"
-    | "headRefOid"
-  >;
-  checks: PullRequestCheck[];
-}): string | null {
-  if (!args.record.last_head_sha || args.record.last_head_sha === args.pr.headRefOid) {
-    return null;
-  }
-
-  return firstPassingCheckEvidence(args.checks) ?? currentHeadConfiguredBotEvidence(args.pr);
 }
 
 export function applyRecoveryEvent(
@@ -754,67 +666,6 @@ export async function reconcileStaleDoneIssueStates(
   );
 }
 
-function isCurrentHeadReviewSignalRequestTimeout(
-  patch: Pick<
-    IssueRunRecord,
-    "copilot_review_timed_out_at" | "copilot_review_timeout_action" | "copilot_review_timeout_reason"
-  >,
-): boolean {
-  return (
-    patch.copilot_review_timed_out_at !== null &&
-    patch.copilot_review_timeout_action === "request_review_comment" &&
-    patch.copilot_review_timeout_reason?.includes("current-head review signal") === true
-  );
-}
-
-function hasOnlyOutdatedConfiguredBotResidue(
-  config: SupervisorConfig,
-  reviewThreads: ReviewThread[],
-): boolean {
-  const unresolvedThreads = reviewThreads.filter((thread) => !thread.isResolved);
-  if (unresolvedThreads.length === 0) {
-    return false;
-  }
-
-  const configuredThreads = configuredBotReviewThreads(config, unresolvedThreads);
-  return (
-    configuredThreads.length === unresolvedThreads.length &&
-    configuredThreads.every(
-      (thread) =>
-        thread.isOutdated &&
-        (latestReviewCommentAuthorIsAllowedBot(config, thread) || operatorAcknowledgedCodexResidue(thread)),
-    )
-  );
-}
-
-function operatorAcknowledgedCodexResidue(thread: ReviewThread): boolean {
-  if (!hasCodexConnectorFindingReviewComment(thread)) {
-    return false;
-  }
-
-  const latestComment = latestReviewComment(thread);
-  if (!latestComment || latestComment.author?.typeName === "Bot") {
-    return false;
-  }
-
-  const normalizedBody = latestComment.body.replace(/\s+/g, " ").trim().toLowerCase();
-  return (
-    /\bresidue acknowledged\b/.test(normalizedBody) ||
-    /\bstale codex connector (?:finding|residue)\b/.test(normalizedBody) ||
-    /\bcovered by the current-head success signal\b/.test(normalizedBody) ||
-    hasSupervisorVerifiedNoSourceChangeStaleAcknowledgement(normalizedBody)
-  );
-}
-
-function hasSupervisorVerifiedNoSourceChangeStaleAcknowledgement(normalizedBody: string): boolean {
-  return (
-    normalizedBody.includes("the supervisor reprocessed this configured-bot finding") &&
-    normalizedBody.includes("classified it as stale") &&
-    normalizedBody.includes("reason=verified_no_source_change_auto_resolve") &&
-    normalizedBody.includes("processed_on_current_head=yes")
-  );
-}
-
 export async function reconcileRecoverableBlockedIssueStates(
   github: Pick<RecoveryGitHubLike, "getPullRequestIfExists" | "getIssue" | "getChecks" | "getUnresolvedReviewThreads">
     & Partial<Pick<RecoveryGitHubLike, "addIssueComment" | "getExternalReviewSurface" | "getIssueComments" | "updateIssueComment">>,
@@ -822,566 +673,19 @@ export async function reconcileRecoverableBlockedIssueStates(
   state: SupervisorStateFile,
   config: SupervisorConfig,
   issues: GitHubIssue[],
-  deps: {
-    shouldAutoRetryHandoffMissing: (record: IssueRunRecord, config: SupervisorConfig) => boolean;
-    inferStateFromPullRequest?: typeof inferStateFromPullRequest;
-    inferFailureContext?: typeof inferFailureContext;
-    blockedReasonForLifecycleState?: typeof blockedReasonForLifecycleState;
-    isOpenPullRequest?: typeof isOpenPullRequest;
-    syncReviewWaitWindow?: typeof syncReviewWaitWindow;
-    syncCopilotReviewRequestObservation?: typeof syncCopilotReviewRequestObservation;
-    syncCopilotReviewTimeoutState?: typeof syncCopilotReviewTimeoutState;
-  },
-  options: {
-    onlyTrackedPrStates?: boolean;
-  } = {},
+  deps: Parameters<typeof reconcileRecoverableBlockedIssueStatesInModule>[5],
+  options: Parameters<typeof reconcileRecoverableBlockedIssueStatesInModule>[6] = {},
 ): Promise<RecoveryEvent[]> {
-  let changed = false;
-  const recoveryEvents: RecoveryEvent[] = [];
-  const issuesByNumber = new Map(issues.map((issue) => [issue.number, issue]));
-  const inferStateFromPullRequestImpl = deps.inferStateFromPullRequest ?? inferStateFromPullRequest;
-  const inferFailureContextImpl = deps.inferFailureContext ?? inferFailureContext;
-  const blockedReasonForLifecycleStateImpl =
-    deps.blockedReasonForLifecycleState ?? blockedReasonForLifecycleState;
-  const isOpenPullRequestImpl = deps.isOpenPullRequest ?? isOpenPullRequest;
-  const syncReviewWaitWindowImpl = deps.syncReviewWaitWindow ?? syncReviewWaitWindow;
-  const syncCopilotReviewRequestObservationImpl =
-    deps.syncCopilotReviewRequestObservation ?? syncCopilotReviewRequestObservation;
-  const syncCopilotReviewTimeoutStateImpl =
-    deps.syncCopilotReviewTimeoutState ?? syncCopilotReviewTimeoutState;
-
-  for (const record of Object.values(state.issues)) {
-    if (record.state !== "blocked") {
-      continue;
-    }
-    if (options.onlyTrackedPrStates && record.pr_number === null) {
-      continue;
-    }
-
-    let issue = issuesByNumber.get(record.issue_number);
-    if (!issue && record.pr_number !== null) {
-      try {
-        issue = await github.getIssue(record.issue_number);
-      } catch {
-        issue = undefined;
-      }
-    }
-
-    if (!issue || issue.state !== "OPEN") {
-      continue;
-    }
-
-    if (deps.shouldAutoRetryHandoffMissing(record, config)) {
-      const recoveryEvent = buildRecoveryEvent(
-        record.issue_number,
-        `stale_state_cleanup: requeued issue #${record.issue_number} after recovering a missing handoff`,
-      );
-      const updated = stateStore.touch(record, {
-        state: "queued",
-        blocked_reason: null,
-        last_error: null,
-        last_failure_kind: null,
-        last_blocker_signature: null,
-        codex_session_id: null,
-        review_wait_started_at: null,
-        review_wait_head_sha: null,
-        copilot_review_requested_observed_at: null,
-        copilot_review_requested_head_sha: null,
-        copilot_review_timed_out_at: null,
-        copilot_review_timeout_action: null,
-        copilot_review_timeout_reason: null,
-        ...applyRecoveryEvent({}, recoveryEvent),
-      });
-      state.issues[String(record.issue_number)] = updated;
-      changed = true;
-      recoveryEvents.push(recoveryEvent);
-      continue;
-    }
-
-    if (shouldReconsiderBlockedNoPrStaleManualStop(record, issue)) {
-      const recoveryEvent = buildRecoveryEvent(
-        record.issue_number,
-        `github_issue_reconsidered: requeued issue #${record.issue_number} after GitHub issue updates arrived following a stale no-PR manual stop`,
-      );
-      const updated = stateStore.touch(record, {
-        state: "queued",
-        blocked_reason: null,
-        last_error: null,
-        last_failure_kind: null,
-        last_failure_context: null,
-        last_blocker_signature: null,
-        last_failure_signature: null,
-        repeated_failure_signature_count: 0,
-        stale_stabilizing_no_pr_recovery_count: 0,
-        codex_session_id: null,
-        ...applyRecoveryEvent({}, recoveryEvent),
-      });
-      state.issues[String(record.issue_number)] = updated;
-      changed = true;
-      recoveryEvents.push(recoveryEvent);
-      continue;
-    }
-
-    if (shouldReconsiderGenericNoPrIssueDefinitionChange(record, issue)) {
-      const recoveryEvent = buildRecoveryEvent(
-        record.issue_number,
-        `github_issue_definition_changed: requeued issue #${record.issue_number} after a material GitHub issue definition change invalidated the stale no-PR ${record.state} state`,
-      );
-      const updated = stateStore.touch(record, {
-        state: "queued",
-        blocked_reason: null,
-        last_error: null,
-        last_failure_kind: null,
-        last_failure_context: null,
-        last_blocker_signature: null,
-        last_failure_signature: null,
-        repeated_failure_signature_count: 0,
-        stale_stabilizing_no_pr_recovery_count: 0,
-        codex_session_id: null,
-        ...issueDefinitionFreshnessPatch(issue),
-        ...applyRecoveryEvent({}, recoveryEvent),
-      });
-      state.issues[String(record.issue_number)] = updated;
-      changed = true;
-      recoveryEvents.push(recoveryEvent);
-      continue;
-    }
-
-    if (
-      record.state === "blocked" &&
-      record.blocked_reason === "handoff_missing" &&
-      record.pr_number !== null
-    ) {
-      const trackedPullRequest = await github.getPullRequestIfExists(record.pr_number, { purpose: "action" });
-      const repairContext = trackedPullRequest
-        ? queuedReadyPromotionPathHygieneRepairContext(record, trackedPullRequest)
-        : null;
-      if (trackedPullRequest && repairContext) {
-        const recoveryEvent = buildTrackedPrResumeRecoveryEvent(
-          record,
-          trackedPullRequest,
-          "repairing_ci",
-          buildRecoveryEvent,
-        );
-        const headAdvanceResetPatch = resetTrackedPrHeadScopedStateOnAdvance(record, trackedPullRequest.headRefOid);
-        const failureSignatureBaseRecord = {
-          ...record,
-          last_failure_signature: null,
-          repeated_failure_signature_count: 0,
-        };
-        const updated = stateStore.touch(record, applyRecoveryEvent({
-          state: "repairing_ci",
-          blocked_reason: null,
-          last_error: truncate(repairContext.summary, 1000),
-          last_failure_kind: null,
-          last_failure_context: repairContext,
-          ...applyFailureSignature(failureSignatureBaseRecord, repairContext),
-          last_blocker_signature: null,
-          pr_number: trackedPullRequest.number,
-          last_head_sha: trackedPullRequest.headRefOid,
-          codex_session_id: null,
-          ...headAdvanceResetPatch,
-        }, recoveryEvent));
-        state.issues[String(record.issue_number)] = updated;
-        changed = true;
-        recoveryEvents.push(recoveryEvent);
-        continue;
-      }
-
-      if (!trackedPullRequest || trackedPullRequest.state !== "OPEN" || trackedPullRequest.mergedAt) {
-        continue;
-      }
-
-      if (mergeConflictDetected(trackedPullRequest)) {
-        const recoveryEvent = buildTrackedPrResumeRecoveryEvent(
-          record,
-          trackedPullRequest,
-          "resolving_conflict",
-          buildRecoveryEvent,
-        );
-        const headAdvanceResetPatch = resetTrackedPrHeadScopedStateOnAdvance(record, trackedPullRequest.headRefOid);
-        const headAdvanced = Object.keys(headAdvanceResetPatch).length > 0;
-        const failureSignatureBaseRecord = headAdvanced
-          ? {
-            ...record,
-            last_failure_signature: null,
-            repeated_failure_signature_count: 0,
-          }
-          : record;
-        const updated = stateStore.touch(record, applyRecoveryEvent({
-          state: "resolving_conflict",
-          blocked_reason: null,
-          last_error: null,
-          last_failure_kind: null,
-          last_failure_context: null,
-          last_blocker_signature: null,
-          ...applyFailureSignature(failureSignatureBaseRecord, null),
-          repeated_blocker_count: 0,
-          repair_attempt_count: 0,
-          timeout_retry_count: 0,
-          blocked_verification_retry_count: 0,
-          codex_session_id: null,
-          pr_number: trackedPullRequest.number,
-          last_head_sha: trackedPullRequest.headRefOid,
-          ...headAdvanceResetPatch,
-        }, recoveryEvent));
-        state.issues[String(record.issue_number)] = updated;
-        changed = true;
-        recoveryEvents.push(recoveryEvent);
-        continue;
-      }
-
-      const checks = await github.getChecks(trackedPullRequest.number);
-      const reviewThreads = await github.getUnresolvedReviewThreads(trackedPullRequest.number);
-      const projection = projectTrackedPrLifecycle({
-        config,
-        record,
-        pr: trackedPullRequest,
-        checks,
-        reviewThreads,
-        inferStateFromPullRequest: inferStateFromPullRequestImpl,
-        blockedReasonForLifecycleState: blockedReasonForLifecycleStateImpl,
-        syncReviewWaitWindow: syncReviewWaitWindowImpl,
-        syncCopilotReviewRequestObservation: syncCopilotReviewRequestObservationImpl,
-        syncCopilotReviewTimeoutState: syncCopilotReviewTimeoutStateImpl,
-      });
-      if (projection.shouldSuppressRecovery) {
-        continue;
-      }
-
-      const externalProgressEvidence = trackedHandoffExternalProgressEvidence({
-        record,
-        pr: trackedPullRequest,
-        checks,
-      });
-      const sameHeadReviewRequestRecovery =
-        externalProgressEvidence === null &&
-        record.last_head_sha === trackedPullRequest.headRefOid &&
-        projection.nextState === "waiting_ci" &&
-        isCurrentHeadReviewSignalRequestTimeout(projection.copilotReviewTimeoutPatch);
-      const isReviewSignalProjectedState =
-        projection.nextState === "addressing_review" || projection.nextState === "waiting_ci";
-      const sameHeadOutdatedConfiguredBotResidueRecovery =
-        externalProgressEvidence === null &&
-        record.last_head_sha === trackedPullRequest.headRefOid &&
-        isReviewSignalProjectedState &&
-        isCurrentHeadReviewSignalRequestTimeout(projection.copilotReviewTimeoutPatch) &&
-        hasOnlyOutdatedConfiguredBotResidue(config, reviewThreads);
-      const sameHeadProviderSuccessRecovery =
-        externalProgressEvidence === null &&
-        record.last_head_sha === trackedPullRequest.headRefOid &&
-        projection.nextState !== "blocked" &&
-        projection.mergeLatencyVisibilityPatch.provider_success_head_sha === trackedPullRequest.headRefOid &&
-        (reviewThreads.filter((thread) => !thread.isResolved).length === 0 ||
-          hasOnlyOutdatedConfiguredBotResidue(config, reviewThreads));
-      if (
-        !externalProgressEvidence &&
-        !sameHeadReviewRequestRecovery &&
-        !sameHeadOutdatedConfiguredBotResidueRecovery &&
-        !sameHeadProviderSuccessRecovery
-      ) {
-        const failureContext = inferFailureContextImpl(config, projection.recordForState, trackedPullRequest, checks, reviewThreads);
-        const updated = await syncTrackedPrPersistentStatusComment({
-          github,
-          stateStore,
-          state,
-          record,
-          pr: trackedPullRequest,
-          checks,
-          reviewThreads,
-          syncJournal: async () => undefined,
-          config,
-          failureContext,
-          summarizeChecks,
-          manualReviewThreadCount: manualReviewThreads(config, reviewThreads).length,
-          skipAutoHandleStaleConfiguredBotReview: true,
-        });
-        if (updated !== record) {
-          state.issues[String(updated.issue_number)] = updated;
-        }
-        continue;
-      }
-
-      const nextState =
-        sameHeadOutdatedConfiguredBotResidueRecovery && !sameHeadProviderSuccessRecovery
-          ? "waiting_ci"
-          : projection.nextState;
-      const nextBlockedReason =
-        sameHeadOutdatedConfiguredBotResidueRecovery && !sameHeadProviderSuccessRecovery
-          ? null
-          : projection.nextBlockedReason;
-      const failureContext =
-        nextState === "blocked"
-          ? inferFailureContextImpl(config, projection.recordForState, trackedPullRequest, checks, reviewThreads)
-          : null;
-      const patch = buildTrackedPrStaleFailureConvergencePatch({
-        record,
-        pr: trackedPullRequest,
-        nextState,
-        failureContext,
-        blockedReason: nextBlockedReason,
-        reviewWaitPatch: projection.reviewWaitPatch,
-        codexConnectorReviewRequestObservationPatch: projection.codexConnectorReviewRequestObservationPatch,
-        copilotReviewRequestObservationPatch: projection.copilotReviewRequestObservationPatch,
-        copilotReviewTimeoutPatch: projection.copilotReviewTimeoutPatch,
-        mergeLatencyVisibilityPatch: projection.mergeLatencyVisibilityPatch,
-      });
-      patch.codex_session_id = null;
-      patch.last_tracked_pr_progress_summary = externalProgressEvidence
-        ? `handoff_missing_recovered=evidence=${externalProgressEvidence}`
-        : `handoff_missing_recovered=same_head_projected_state=${nextState}`;
-
-      const recoveryReason = externalProgressEvidence
-        ? `tracked_pr_handoff_missing_external_progress: resumed issue #${record.issue_number} from blocked to ${nextState} after tracked PR #${trackedPullRequest.number} advanced from ${record.last_head_sha} to ${trackedPullRequest.headRefOid} with evidence=${externalProgressEvidence}`
-        : `tracked_pr_handoff_missing_same_head_recovered: resumed issue #${record.issue_number} from blocked to ${nextState} using fresh tracked PR #${trackedPullRequest.number} facts at head ${trackedPullRequest.headRefOid}`;
-      const recoveryEvent = buildRecoveryEvent(record.issue_number, recoveryReason);
-      const updated = stateStore.touch(record, applyRecoveryEvent(patch, recoveryEvent));
-      state.issues[String(record.issue_number)] = updated;
-      changed = true;
-      recoveryEvents.push(recoveryEvent);
-      continue;
-    }
-
-    if (
-      record.pr_number !== null &&
-      (
-        record.blocked_reason === null ||
-        record.blocked_reason === "manual_review" ||
-        record.blocked_reason === "verification" ||
-        shouldReconcileTrackedPrStaleReviewBot(record, config)
-      )
-    ) {
-      const trackedPullRequest = await github.getPullRequestIfExists(record.pr_number, { purpose: "action" });
-      if (!trackedPullRequest || !isOpenPullRequestImpl(trackedPullRequest)) {
-        continue;
-      }
-
-      const checks = await github.getChecks(trackedPullRequest.number);
-      const reviewThreads = await github.getUnresolvedReviewThreads(trackedPullRequest.number);
-      const projection = projectTrackedPrLifecycle({
-        config,
-        record,
-        pr: trackedPullRequest,
-        checks,
-        reviewThreads,
-        inferStateFromPullRequest: inferStateFromPullRequestImpl,
-        blockedReasonForLifecycleState: blockedReasonForLifecycleStateImpl,
-        syncReviewWaitWindow: syncReviewWaitWindowImpl,
-        syncCopilotReviewRequestObservation: syncCopilotReviewRequestObservationImpl,
-        syncCopilotReviewTimeoutState: syncCopilotReviewTimeoutStateImpl,
-      });
-      const nextState = projection.nextState;
-      if (projection.shouldSuppressRecovery) {
-        continue;
-      }
-      const staleLocalManualReviewResidueRecovery =
-        record.blocked_reason === "manual_review" &&
-        record.last_tracked_pr_repeat_failure_decision === "stop_no_progress" &&
-        nextState !== "blocked" &&
-        hasOnlyOutdatedConfiguredBotResidue(config, reviewThreads);
-      const recoverySuppression = staleLocalManualReviewResidueRecovery
-        ? { shouldSuppress: false, progressSummary: "stale_local_blocker_recovered=outdated_configured_bot_residue" }
-        : suppressSameHeadNoProgressReviewThreadRecovery(
-          record,
-          trackedPullRequest,
-          reviewThreads,
-          nextState,
-        );
-      if (recoverySuppression.shouldSuppress) {
-        const suppressionPatch: Partial<IssueRunRecord> = {
-          last_tracked_pr_progress_summary: recoverySuppression.progressSummary,
-        };
-        if (needsRecordUpdate(record, suppressionPatch)) {
-          const updated = stateStore.touch(record, suppressionPatch);
-          state.issues[String(record.issue_number)] = updated;
-          changed = true;
-        }
-        continue;
-      }
-
-      const inferredFailureContext =
-        nextState === "blocked"
-        || (
-          nextState === "draft_pr"
-          && record.blocked_reason === "verification"
-          && trackedPullRequest.isDraft
-        )
-          ? inferFailureContextImpl(config, projection.recordForState, trackedPullRequest, checks, reviewThreads)
-          : null;
-      const preserveDraftReadyPromotionBlocker =
-        nextState === "draft_pr"
-        && record.blocked_reason === "verification"
-        && trackedPullRequest.isDraft
-        && (
-          inferredFailureContext !== null ||
-          hasFreshTrackedPrReadyPromotionBlockerEvidence(record, trackedPullRequest)
-        );
-      const failureContext =
-        inferredFailureContext
-        ?? (preserveDraftReadyPromotionBlocker ? record.last_failure_context : null);
-      const nextBlockedReason = preserveDraftReadyPromotionBlocker
-        ? "verification"
-        : projection.nextBlockedReason;
-
-      if (nextState === "blocked" || preserveDraftReadyPromotionBlocker) {
-        const headAdvanceResetPatch = resetTrackedPrHeadScopedStateOnAdvance(record, trackedPullRequest.headRefOid);
-        const headAdvanced = Object.keys(headAdvanceResetPatch).length > 0;
-        const blockerSemanticsChanged =
-          headAdvanced
-          || nextBlockedReason !== record.blocked_reason
-          || (failureContext?.signature ?? null) !== record.last_failure_signature;
-        const failureSignatureBaseRecord = headAdvanced
-          ? {
-            ...record,
-            last_failure_signature: null,
-            repeated_failure_signature_count: 0,
-          }
-          : record;
-        const failureSignaturePatch =
-          preserveDraftReadyPromotionBlocker && !blockerSemanticsChanged
-            ? {
-              last_failure_signature: record.last_failure_signature,
-              repeated_failure_signature_count: record.repeated_failure_signature_count,
-            }
-            : applyFailureSignature(failureSignatureBaseRecord, failureContext);
-        const blockedPatch: Partial<IssueRunRecord> = {
-          state: "blocked",
-          last_error: failureContext ? truncate(failureContext.summary, 1000) : null,
-          last_failure_kind: null,
-          last_failure_context: failureContext,
-          last_blocker_signature: null,
-          ...failureSignaturePatch,
-          blocked_reason: nextBlockedReason,
-          pr_number: trackedPullRequest.number,
-          last_head_sha: trackedPullRequest.headRefOid,
-          ...headAdvanceResetPatch,
-          ...projection.reviewWaitPatch,
-          ...projection.copilotReviewRequestObservationPatch,
-          ...projection.copilotReviewTimeoutPatch,
-        };
-        const nextPatch = blockerSemanticsChanged
-          ? {
-            ...blockedPatch,
-            repeated_blocker_count: 0,
-            repair_attempt_count: 0,
-            timeout_retry_count: 0,
-            blocked_verification_retry_count: 0,
-          }
-          : blockedPatch;
-
-        if (needsRecordUpdate(record, nextPatch)) {
-          const updated = stateStore.touch(record, nextPatch);
-          state.issues[String(record.issue_number)] = updated;
-          changed = true;
-        }
-        continue;
-      }
-
-      const patch = buildTrackedPrStaleFailureConvergencePatch({
-        record,
-        pr: trackedPullRequest,
-        nextState,
-        failureContext,
-        blockedReason: nextBlockedReason,
-        reviewWaitPatch: projection.reviewWaitPatch,
-        codexConnectorReviewRequestObservationPatch: projection.codexConnectorReviewRequestObservationPatch,
-        copilotReviewRequestObservationPatch: projection.copilotReviewRequestObservationPatch,
-        copilotReviewTimeoutPatch: projection.copilotReviewTimeoutPatch,
-        mergeLatencyVisibilityPatch: projection.mergeLatencyVisibilityPatch,
-      });
-      if (!staleLocalManualReviewResidueRecovery && recoverySuppression.progressSummary !== null) {
-        patch.last_tracked_pr_progress_summary = recoverySuppression.progressSummary;
-      }
-      if (staleLocalManualReviewResidueRecovery) {
-        patch.last_tracked_pr_progress_snapshot = null;
-        patch.last_tracked_pr_progress_summary = null;
-        patch.last_tracked_pr_repeat_failure_decision = null;
-        patch.processed_review_thread_ids = [];
-        patch.processed_review_thread_fingerprints = [];
-      }
-      const recoveryEvent = staleLocalManualReviewResidueRecovery
-        ? buildRecoveryEvent(
-          record.issue_number,
-          `tracked_pr_stale_local_blocker_recovered: resumed issue #${record.issue_number} from blocked to ${nextState} after stale manual-review metadata was superseded by tracked PR #${trackedPullRequest.number} facts at head ${trackedPullRequest.headRefOid}`,
-        )
-        : buildTrackedPrResumeRecoveryEvent(
-          record,
-          trackedPullRequest,
-          nextState,
-          buildRecoveryEvent,
-        );
-      const updated = stateStore.touch(record, applyRecoveryEvent(patch, recoveryEvent));
-      state.issues[String(record.issue_number)] = updated;
-      changed = true;
-      recoveryEvents.push(recoveryEvent);
-      continue;
-    }
-
-    if (record.state === "blocked" && record.blocked_reason === "requirements") {
-      if (!hasAvailableIssueLabels(issue)) {
-        continue;
-      }
-
-      const readiness = lintExecutionReadyIssueBody(issue);
-      if (!readiness.isExecutionReady) {
-        continue;
-      }
-
-      const recoveryEvent = buildRecoveryEvent(
-        record.issue_number,
-        `requirements_recovered: requeued issue #${record.issue_number} after execution-ready metadata was added`,
-      );
-      const updated = stateStore.touch(record, {
-        state: "queued",
-        blocked_reason: null,
-        last_error: null,
-        last_failure_kind: null,
-        last_failure_context: null,
-        last_blocker_signature: null,
-        last_failure_signature: null,
-        repeated_failure_signature_count: 0,
-        ...applyRecoveryEvent({}, recoveryEvent),
-      });
-      state.issues[String(record.issue_number)] = updated;
-      changed = true;
-      await clearRequirementsBlockerIssueComment(github, record.issue_number, issue.updatedAt);
-      recoveryEvents.push(recoveryEvent);
-      continue;
-    }
-
-    if (record.state === "blocked" && record.blocked_reason === "clarification") {
-      if (findHighRiskBlockingAmbiguity(issue)) {
-        continue;
-      }
-
-      const recoveryEvent = buildRecoveryEvent(
-        record.issue_number,
-        `clarification_recovered: requeued issue #${record.issue_number} after blocking ambiguity was resolved`,
-      );
-      const updated = stateStore.touch(record, {
-        state: "queued",
-        blocked_reason: null,
-        last_error: null,
-        last_failure_kind: null,
-        last_failure_context: null,
-        last_blocker_signature: null,
-        last_failure_signature: null,
-        repeated_failure_signature_count: 0,
-        ...applyRecoveryEvent({}, recoveryEvent),
-      });
-      state.issues[String(record.issue_number)] = updated;
-      changed = true;
-      recoveryEvents.push(recoveryEvent);
-    }
-  }
-
-  if (changed) {
-    await stateStore.save(state);
-  }
-
-  return recoveryEvents;
+  return reconcileRecoverableBlockedIssueStatesInModule(
+    github,
+    stateStore,
+    state,
+    config,
+    issues,
+    deps,
+    options,
+  );
 }
-
 export async function reconcileParentEpicClosures(
   github: RecoveryGitHubLike,
   stateStore: StateStoreLike,
